Make DeepSpeed main branch! (#3)
